refactor the routes for Flask
